Changed file structure.
